+++ CHANGELOG.md
@@ -37,8 +37,9 @@ BREAKING CHANGES
   * Add REST endpoint to unrevoke a validator previously revoked for downtime
   * Add REST endpoint to retrieve liveness signing information for a validator
 * [types] renamed rational.Evaluate to rational.Round{Int64, Int}
-* [stake] most index keys nolonger hold a value - inputs are rearranged to form the desired key
+* [x/stake] most index keys nolonger hold a value - inputs are rearranged to form the desired key
 * [lcd] Switch key creation output to return bech32
+* [x/stake] store-value for delegation, validator, ubd, and red do not hold duplicate information contained store-key
 
 FEATURES
 * [gaiacli] You can now attach a simple text-only memo to any transaction, with the `--memo` flag